Fixed a display bug of multiple bars
In fact the fix seems to be quiet simple :). Fixes bugs when displaying multiple bars when running nested loops or providing `position` argument
The sad thing is that the fix breaks the very easiest way of using tqdm: just `for i in trange`, because after the loop ends, any following print (there is also a way to remove the `fp_write` line completely, but it results into unexpected behaviour with multiple bars runned one after another without any special args (the second bar will override the first one). I find this worse, than the current version)
Anyway, this is just the first step, I am going to fix all of the above