Added automatic lines termination after bars closing
Whenever a group of bars (e. g. from nested loops, parallel execution, etc) finishes (which means all the bars are closed), cursor moves below the printed bars and a user program is able to write to a file.
However, this still does not work for one (extremely special as for me) case: when two bars with `position`s specified are created, but one of them finishes before the other one starts (for example, if you do something like `for i in trange(..., pos=1): pass; for i in trange(..., pos=2): pass;`. In such case the output will still be broken (however it will be easier to read, than before this fix).
But, anyway, I'm going to improve the behaviour described above either